Rename indexer to be more explicit
