+++ packages/documentation/src/utils/index.ts
@@ -1,7 +1,7 @@
 export * from './bomb-args';
 export * from './component-properties';
 export * from './get-attributes';
-export * from './indexer';
+export * from './storybook-indexer';
 export * from './map-classes';
 export * from './spread-args';
 export * from './figma-link';
